Don't attempt to apply solution from salvage.
Match the behavior of typeCheckExpressionImpl which simply calls solve
and applySolution but never salvage.
