+++ test/Parse/omit_return_foundation.swift
@@ -1,0 +1,20 @@
+// RUN: %target-swift-frontend %s -typecheck -verify
+
+import Foundation
+
+class DynamicSubscriptClass {
+  @objc subscript (i : Int) -> String { "howdy" }
+}
+
+func ff_implicitDynamicSubscript(_ c: DynamicSubscriptClass) -> String {
+    c[13]
+}
+
+func ff_implicitObjcSelectorExpr() -> Selector {
+    #selector(NSArray.object(at:))
+}
+
+func ff_implicitKeyPathExpr() -> String {
+    #keyPath(NSArray.count)
+}
+
